PowerOfTen function return type changed for floating point number result option

+++ c/include/ws1.h
@@ -4,7 +4,7 @@
 
 void PrintHello();
 
-int PowerOfTen(int n);
+double PowerOfTen(int n);
 
 int FlipOrder(int n);
